#9 move augment data to its own directory
