+++ hygia/__init__.py
@@ -5,7 +5,7 @@ from hygia.data_pipeline.feature_engineering.word_embedding import (WordEmbeddin
 from hygia.data_pipeline.model.random_forest import (RandomForestModel)
 from hygia.data_pipeline.pre_process_data.pre_process_data import (PreProcessData)
 from hygia.data_pipeline.annotate_data.annotate_data import (AnnotateData)
-from hygia.data_pipeline.pre_process_data.augment_data import (AugmentData)
+from hygia.data_pipeline.augment_data.augment_data import (AugmentData)
 from hygia.main import (run_with_config)
 
 __all__ = [
