+++ tests/data_pipeline/augment_data/test_augment_data.py
@@ -1,6 +1,6 @@
 import pandas as pd
 import pytest
-from hygia.data_pipeline.pre_process_data.augment_data import AugmentData
+from hygia.data_pipeline.augment_data.augment_data import AugmentData
 
 @pytest.mark.parametrize("zipcode, expected", [
     ('20000', True),
